+++ whalesong/managers/wap.py
@@ -11,14 +11,14 @@ class Wap(FastDynamicModel):
     pass
 
 
-class WapManager(BaseModelManager):
+class WapManager(BaseModelManager[Wap]):
     """
     Entry point to request data to phone.
     """
 
     MODEL_CLASS = Wap
 
-    def query_exist(self, contact_id: str) -> Result:
+    def query_exist(self, contact_id: str) -> Result[bool]:
         """
         Check whether a contact identifier exists on Whatsapp or not.